+++ lib/live_view_apps_web/components/custom_components.ex
@@ -1,0 +1,23 @@
+defmodule LiveViewAppsWeb.CustomComponents do
+  use Phoenix.Component
+
+  attr :expiration, :integer, default: 24
+  slot :legal
+  slot :inner_block, required: true
+
+  def promo(assigns) do
+    ~H"""
+    <div class="promo">
+      <div class="deal">
+        <%= render_slot(@inner_block) %>
+      </div>
+      <div class="expiration">
+        Deal expires on <%= @expiration %> hours!
+      </div>
+      <div class="legal">
+        <%= render_slot(@legal) %>
+      </div>
+    </div>
+    """
+  end
+end
