+++ lib/live_view_apps_web/live/boats_live.ex
@@ -2,6 +2,7 @@ defmodule LiveViewAppsWeb.BoatsLive do
   use LiveViewAppsWeb, :live_view
 
   alias LiveViewApps.Boats
+  import LiveViewAppsWeb.CustomComponents, only: [promo: 1]
 
   def mount(_params, _session, socket) do
     socket =
@@ -25,50 +26,13 @@ defmodule LiveViewAppsWeb.BoatsLive do
           <Heroicons.exclamation_circle /> Limit 1 per party
         </:legal>
       </.promo>
-      <form phx-change="filter">
-        <div class="filters">
-          <select name="type">
-            <%= Phoenix.HTML.Form.options_for_select(
-              type_options(),
-              @filter.type
-            ) %>
-          </select>
-          <div class="prices">
-            <%= for price <- ["$", "$$", "$$$"] do %>
-              <input
-                type="checkbox"
-                name="prices[]"
-                value={price}
-                id={price}
-                checked={price in @filter.prices}
-              />
-              <label for={price}><%= price %></label>
-            <% end %>
-            <input type="hidden" name="prices[]" value="" />
-          </div>
-        </div>
-      </form>
+      <.filter_form filter={@filter} />
       <div class="boats">
-        <div :for={boat <- @boats} class="boat">
-          <img src={boat.image} />
-          <div class="content">
-            <div class="model">
-              <%= boat.model %>
-            </div>
-            <div class="details">
-              <span class="price">
-                <%= boat.price %>
-              </span>
-              <span class="type">
-                <%= boat.type %>
-              </span>
-            </div>
-          </div>
-        </div>
+        <.boat :for={boat <- @boats} boat={boat} />
       </div>
       <.promo expiration={2}>
         <p>
-          Save 25% on your first rental
+          Last 3 boats available!
         </p>
         <:legal>
           <Heroicons.exclamation_circle /> Limit 1 per party
@@ -78,17 +42,54 @@ defmodule LiveViewAppsWeb.BoatsLive do
     """
   end
 
-  def promo(assigns) do
+  attr :filter, :map, required: true
+
+  def filter_form(assigns) do
     ~H"""
-    <div class="promo">
-      <div class="deal">
-        <%= render_slot(@inner_block) %>
+    <form phx-change="filter">
+      <div class="filters">
+        <select name="type">
+          <%= Phoenix.HTML.Form.options_for_select(
+            type_options(),
+            @filter.type
+          ) %>
+        </select>
+        <div class="prices">
+          <%= for price <- ["$", "$$", "$$$"] do %>
+            <input
+              type="checkbox"
+              name="prices[]"
+              value={price}
+              id={price}
+              checked={price in @filter.prices}
+            />
+            <label for={price}><%= price %></label>
+          <% end %>
+          <input type="hidden" name="prices[]" value="" />
+        </div>
       </div>
-      <div class="expiration">
-        Deal expires on <%= @expiration %> hours!
-      </div>
-      <div class="legal">
-        <%= render_slot(@legal) %>
+    </form>
+    """
+  end
+
+  attr :boat, LiveViewApps.Boats.Boat, required: true
+
+  def boat(assigns) do
+    ~H"""
+    <div class="boat">
+      <img src={@boat.image} />
+      <div class="content">
+        <div class="model">
+          <%= @boat.model %>
+        </div>
+        <div class="details">
+          <span class="price">
+            <%= @boat.price %>
+          </span>
+          <span class="type">
+            <%= @boat.type %>
+          </span>
+        </div>
       </div>
     </div>
     """
